refactor(auth): tidy confirmation-window config and document within() scope

+++ config/auth.php
@@ -154,6 +154,6 @@ return [
 
     'password_timeout' => env('AUTH_PASSWORD_TIMEOUT', 10800),
 
-    'confirmation_window' => (int) env('AUTH_CONFIRMATION_WINDOW', 900),
+    'confirmation_window' => env('AUTH_CONFIRMATION_WINDOW', 900),
 
 ];
